Move inbox/outbox classes to the right namespace
