+++ framework/src/Volo.Abp.EventBus/Volo/Abp/EventBus/Distributed/InboxProcessor.cs
@@ -7,12 +7,11 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 using Volo.Abp.DependencyInjection;
 using Volo.Abp.DistributedLocking;
-using Volo.Abp.EventBus.Distributed;
 using Volo.Abp.Threading;
 using Volo.Abp.Timing;
 using Volo.Abp.Uow;
 
-namespace Volo.Abp.EventBus.Boxes;
+namespace Volo.Abp.EventBus.Distributed;
 
 public class InboxProcessor : IInboxProcessor, ITransientDependency
 {
@@ -110,7 +109,7 @@ public class InboxProcessor : IInboxProcessor, ITransientDependency
 
                             await Inbox.MarkAsProcessedAsync(waitingEvent.Id);
 
-                            await uow.CompleteAsync();
+                            await uow.CompleteAsync(StoppingToken);
                         }
 
                         Logger.LogInformation($"Processed the incoming event with id = {waitingEvent.Id:N}");
